+++ front/save_config.php
@@ -2,7 +2,7 @@
 /*
 -------------------------------------------------------------------------
 MailAnalyzer plugin for GLPI
-Copyright (C) 2011-2025 by Raynet SAS a company of A.Raymond Network.
+Copyright (C) 2011-2026 by Raynet SAS a company of A.Raymond Network.
 -------------------------------------------------------------------------
 LICENSE: GPLv2+
 --------------------------------------------------------------------------